backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
# pinecone-client==6.0.0  # Removed as we're using Supabase with pgvector instead
supabase==2.13.0  # Updated to match installed version
openai==1.66.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.2.1
pytest==8.0.0
httpx==0.28.1 